+++ integrations/sqldelight/src/jvmTest/kotlin/com/powersync/integrations/sqldelight/SqlDelightTest.jvm.kt
@@ -1,5 +1,0 @@
-package com.powersync.integrations.sqldelight
-
-import com.powersync.DatabaseDriverFactory
-
-actual fun databaseDriverFactory(): DatabaseDriverFactory = DatabaseDriverFactory()